chore: rename tests.test.tsx
